first round of refactoring
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/explicit_cegar@4161 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/explicit/ExplicitPrecision.java
@@ -36,9 +36,9 @@ public class ExplicitPrecision implements Precision {
 
   final Set<String> whiteList;
 
-  public Map<CFAEdge, Map<String, Integer>> facts = new HashMap<CFAEdge, Map<String, Integer>>();
+  public Map<CFAEdge, Map<String, Long>> facts = new HashMap<CFAEdge, Map<String, Long>>();
 
-  public void setFacts(Map<CFAEdge, Map<String, Integer>> f)
+  public void setFacts(Map<CFAEdge, Map<String, Long>> f)
   {
     facts = f;
   }
